distrivalto-back/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
